modificated model
