+++ app/app.js
@@ -2,11 +2,11 @@ require('dotenv').config();
 
 const express = require('express');
 
-const routerAppoinments = require('./appoinments/router');
-const routerClients = require('./clients/router');
+// const routerAppoinments = require('./appoinments/router');
+// const routerClients = require('./clients/router');
 
 app = express();
-app.use("/appoinments", routerAppoinments);
-app.use("/client", routerClients);
+// app.use("/appoinments", routerAppoinments);
+// app.use("/client", routerClients);
 
 app.listen(process.env.PORT, () => console.log('El servidor esta levantado en', process.env.PORT));
