+++ app/appoinments/controller.js
@@ -1,0 +1,2 @@
+// const express = require('express');
+// const { AppointmentsModel } = require('./appointmentsmodel')
