+++ app/appoinments/router.js
@@ -1,11 +1,11 @@
-const router = require('express').Router();
+// const router = require('express').Router();
 
-const controller = require('./controller');
+// const controller = require('./controller');
 
-router.get('/:id', controller);
-router.get('/', controller);
-router.post('/', controller);
-router.put('/:id', controller);
-router.delete('/id', controller);
+// // router.get('/:id', controller);
+// // router.get('/', controller);
+// // router.post('/', controller);
+// // router.put('/:id', controller);
+// // router.delete('/id', controller);
 
-module.exports = router;
+// module.exports = router;
