+++ app/clients/controller.js
@@ -1,0 +1,13 @@
+// const express = require('express');
+// const { AppointmentsModel } = require('./appointmentsmodel')
+
+// const createUser = (req, res) => {
+//     const user = AppointmentsModel.create(req.body);
+//     user.save();
+//     console.log(user);
+
+// };
+
+// module.exports = {
+//     createUser
+// }
