+++ app/clients/router.js
@@ -1,0 +1,11 @@
+const router = require('express').Router();
+
+const controller = require('./controller');
+
+// // router.get('/:id', controller);
+router.get('/', controller);
+// // router.post('/', controller);
+// // router.put('/:id', controller);
+// // router.delete('/id', controller);
+
+module.exports = router;
