+++ models/appointments.js
@@ -1,0 +1,24 @@
+'use strict';
+const {
+    Model
+} = require('sequelize');
+module.exports = (sequelize, DataTypes) => {
+    class AppointmentModel extends Model {
+        /**
+         * Helper method for defining associations.
+         * This method is not a part of Sequelize lifecycle.
+         * The `models/index` file will call this method automatically.
+         */
+        static associate(models) {
+            // define association here
+        }
+    };
+    AppointmentModel.init({
+        dateNumber: DataTypes.INTEGER,
+        schedule: DataTypes.STRING
+    }, {
+        sequelize,
+        modelName: 'AppointmentModel',
+    });
+    return AppointmentModel;
+};
